requirements.txt
tensorflow==1.15
matplotlib
pillow
scikit-learn
scikit-image
opencv-python

